requirements.txt
Flask
google-generativeai
sentence-transformers
scikit-learn
PyPDF2
python-docx